Moved service accounts snippets to iam/snippets
