+++ iam/snippets/pom.xml
@@ -75,5 +75,15 @@
       <scope>test</scope>
       <version>4.13.2</version>
     </dependency>
+    <dependency>
+      <groupId>com.google.apis</groupId>
+      <artifactId>google-api-services-iam</artifactId>
+      <scope>compile</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.google.apis</groupId>
+      <artifactId>google-api-services-iamcredentials</artifactId>
+      <version>v1-rev20211203-2.0.0</version>
+    </dependency>
   </dependencies>
 </project>
