+++ iam/snippets/src/main/java/EnableServiceAccount.java
@@ -1,4 +1,4 @@
-/* Copyright 2022 Google LLC
+/* Copyright 2019 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,45 +13,36 @@
  * limitations under the License.
  */
 
-package iam.snippets;
-
-// [START iam_create_key]
-
+// [START iam_enable_service_account]
 import com.google.cloud.iam.admin.v1.IAMClient;
-import com.google.gson.Gson;
-import com.google.iam.admin.v1.CreateServiceAccountKeyRequest;
-import com.google.iam.admin.v1.ServiceAccountKey;
+import com.google.iam.admin.v1.EnableServiceAccountRequest;
 import java.io.IOException;
 
-public class CreateServiceAccountKey {
+
+public class EnableServiceAccount {
 
   public static void main(String[] args) throws IOException {
     // TODO(Developer): Replace the below variables before running.
     String projectId = "your-project-id";
     String serviceAccountName = "your-service-account-name";
 
-    ServiceAccountKey key = createKey(projectId, serviceAccountName);
-    Gson gson = new Gson();
-
-    System.out.println("Service account key: " + gson.toJson(key));
+    enableServiceAccount(projectId, serviceAccountName);
   }
 
-  // Creates a key for a service account.
-  public static ServiceAccountKey createKey(String projectId, String accountName)
+  // Enables a service account.
+  public static void enableServiceAccount(String projectId, String accountName)
           throws IOException {
     String email = String.format("%s@%s.iam.gserviceaccount.com", accountName, projectId);
 
     // Initialize client that will be used to send requests.
     // This client only needs to be created once, and can be reused for multiple requests.
     try (IAMClient iamClient = IAMClient.create()) {
-      CreateServiceAccountKeyRequest req = CreateServiceAccountKeyRequest.newBuilder()
+      iamClient.enableServiceAccount(EnableServiceAccountRequest.newBuilder()
               .setName(String.format("projects/%s/serviceAccounts/%s", projectId, email))
-              .build();
-      ServiceAccountKey createdKey = iamClient.createServiceAccountKey(req);
-      System.out.println("Key created successfully");
+              .build());
 
-      return createdKey;
+      System.out.println("Enabled service account: " + email);
     }
   }
 }
-// [END iam_create_key]
+// [END iam_enable_service_account]
